refactor(client->app/createfundingproposal): refactoring & adding react query & husky
