+++ packages/client/package.json
@@ -20,6 +20,7 @@
         "react": "^17.0.2",
         "react-bootstrap": "^2.2.2",
         "react-dom": "^17.0.2",
+        "react-query": "^3.34.19",
         "react-router-dom": "^6.2.2",
         "react-scripts": "5.0.0",
         "stream-browserify": "^3.0.0",
@@ -30,7 +31,9 @@
         "start": "react-app-rewired start",
         "build": "react-app-rewired build",
         "test": "react-scripts test",
-        "eject": "react-scripts eject"
+        "eject": "react-scripts eject",
+        "commit": "git-cz",
+        "precommit": "echo 'client precommit running'"
     },
     "eslintConfig": {
         "extends": [
