+++ packages/client/src/queries/index.ts
@@ -1,0 +1,6 @@
+import { getFundingProposal } from "./getFundingProposal";
+export const Queries = {
+  getFundingProposal,
+};
+
+export const Mutations = {};
