+++ packages/client/src/types/index.ts
@@ -1,0 +1,5 @@
+export interface IGetFundingProposalData {
+  contractAddress: string;
+  grants: [grantee: string, amount: number][];
+  memo: string;
+}
